refactor validate_privilege
