+++ src/query/service/src/sessions/session.rs
@@ -40,6 +40,7 @@ use crate::sessions::SessionManager;
 use crate::sessions::SessionStatus;
 use crate::sessions::SessionType;
 
+use super::session_privilege_mgr::SessionPrivilegeManager;
 use super::session_privilege_mgr::SessionPrivilegeManagerImpl;
 
 pub struct Session {
@@ -250,20 +251,7 @@ impl Session {
     // On executing SET ROLE, the role have to be one of the available roles.
     #[async_backtrace::framed]
     pub async fn get_all_available_roles(self: &Arc<Self>) -> Result<Vec<RoleInfo>> {
-        let roles = match self.session_ctx.get_auth_role() {
-            Some(auth_role) => vec![auth_role],
-            None => {
-                let current_user = self.get_current_user()?;
-                current_user.grants.roles()
-            }
-        };
-
-        let tenant = self.get_current_tenant();
-        let mut related_roles = RoleCacheManager::instance()
-            .find_related_roles(&tenant, &roles)
-            .await?;
-        related_roles.sort_by(|a, b| a.name.cmp(&b.name));
-        Ok(related_roles)
+        todo!()
     }
 
     #[async_backtrace::framed]
@@ -287,51 +275,8 @@ impl Session {
         if matches!(self.get_type(), SessionType::Local) {
             return Ok(());
         }
-
-        // 1. check user's privilege set
-        let current_user = self.get_current_user()?;
-        let user_verified = current_user
-            .grants
-            .verify_privilege(object, privilege.clone());
-        if user_verified {
-            return Ok(());
-        }
-
-        // 2. check the user's roles' privilege set
-        self.ensure_current_role().await?;
-        let available_roles = self.get_all_available_roles().await?;
-        let role_verified = &available_roles
-            .iter()
-            .any(|r| r.grants.verify_privilege(object, privilege.clone()));
-        if *role_verified {
-            return Ok(());
-        }
-
-        if verify_ownership {
-            let object_owner = self.get_object_owner(object).await?;
-            if let Some(owner) = object_owner.as_ref() {
-                for role in &available_roles {
-                    if *role.identity() == *owner.identity() {
-                        return Ok(());
-                    }
-                }
-            }
-        }
-
-        let roles_name = available_roles
-            .iter()
-            .map(|r| r.name.clone())
-            .collect::<Vec<_>>()
-            .join(",");
-
-        Err(ErrorCode::PermissionDenied(format!(
-            "Permission denied, privilege {:?} is required on {} for user {} with roles [{}]",
-            privilege.clone(),
-            object,
-            &current_user.identity(),
-            roles_name,
-        )))
-    }
+        self.privilege_mgr.validate_privilege(object, privilege, verify_ownership).await
+   }
 
     pub fn get_settings(self: &Arc<Self>) -> Arc<Settings> {
         self.session_ctx.get_settings()
